requirements.txt
PyYAML>=6.0.1
anthropic>=0.34.0
colorama>=0.4.6
groq>=0.9.0
jsonschema>=4.21.1
jsonschema>=4.22.0
openai>=1.35.0
openai>=1.40.0,<2
openapi-spec-validator>=0.7.1
prance>=23.6.21
python-dotenv>=1.0.1
pyyaml>=6.0.1
requests>=2.31.0
tqdm>=4.66.4
